refactor: project card padding
