+++ frontend/src/component/project/NewProjectCard/NewProjectCard.styles.ts
@@ -3,6 +3,7 @@ import { Card, Box } from '@mui/material';
 import Delete from '@mui/icons-material/Delete';
 import Edit from '@mui/icons-material/Edit';
 import { flexRow } from 'themes/themeStyles';
+import { ReactComponent as ProjectIcon } from 'assets/icons/projectIconSmall.svg';
 
 export const StyledProjectCard = styled(Card)(({ theme }) => ({
     display: 'flex',
@@ -65,6 +66,7 @@ export const StyledDivInfo = styled('div')(({ theme }) => ({
     display: 'flex',
     justifyContent: 'space-between',
     fontSize: theme.fontSizes.smallerBody,
+    padding: theme.spacing(0, 1),
 }));
 
 export const StyledDivInfoContainer = styled('div')(() => ({
@@ -75,4 +77,15 @@ export const StyledParagraphInfo = styled('p')(({ theme }) => ({
     color: theme.palette.primary.dark,
     fontWeight: 'bold',
     fontSize: theme.typography.body1.fontSize,
+}));
+
+export const StyledProjectIcon = styled(ProjectIcon)(({ theme }) => ({
+    color: theme.palette.primary.main,
+}));
+
+export const StyledIconBox = styled(Box)(({ theme }) => ({
+    display: 'flex',
+    justifyContent: 'center',
+    padding: theme.spacing(0.5, 0.5, 0.5, 0),
+    marginRight: theme.spacing(2),
 }));
